requirements.txt

# external requirements

pandas
numpy
datetime
pyspark
python-dotenv
dbutils
